+++ dev-test/star-wars/__generated__/HeroTypeDependentAliasedField.tsx
@@ -2,7 +2,7 @@ import * as Types from '../types.d';
 
 import { gql } from '@apollo/client';
 import * as Apollo from '@apollo/client';
-const defaultOptions = {};
+const defaultOptions = {} as const;
 export type HeroTypeDependentAliasedFieldQueryVariables = Types.Exact<{
   episode?: Types.InputMaybe<Types.Episode>;
 }>;