+++ packages/plugins/other/fragment-matcher/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @graphql-codegen/fragment-matcher
 
+## 3.2.1
+
+### Patch Changes
+
+- 8643b3bf3: Add GraphQL 16 as a peerDependency
+- 6002feb3d: Fix exports in package.json files for react-native projects
+- Updated dependencies [6002feb3d]
+  - @graphql-codegen/plugin-helpers@2.3.2
+
 ## 3.2.0
 
 ### Minor Changes