+++ packages/plugins/typescript/generic-sdk/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @graphql-codegen/typescript-generic-sdk
 
+## 2.3.1
+
+### Patch Changes
+
+- 6002feb3d: Fix exports in package.json files for react-native projects
+- Updated dependencies [6002feb3d]
+  - @graphql-codegen/visitor-plugin-common@2.5.2
+  - @graphql-codegen/plugin-helpers@2.3.2
+
 ## 2.3.0
 
 ### Minor Changes