+++ packages/plugins/typescript/graphql-request/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @graphql-codegen/typescript-graphql-request
 
+## 4.3.3
+
+### Patch Changes
+
+- 6002feb3d: Fix exports in package.json files for react-native projects
+- Updated dependencies [6002feb3d]
+  - @graphql-codegen/visitor-plugin-common@2.5.2
+  - @graphql-codegen/plugin-helpers@2.3.2
+
 ## 4.3.2
 
 ### Patch Changes