+++ packages/plugins/typescript/jit-sdk/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @graphql-codegen/typescript-jit-sdk
 
+## 2.0.0
+
+### Major Changes
+
+- 170e43474: feat(jit-sdk): ability to provide compiler options while creating the sdk
+
+## 1.2.1
+
+### Patch Changes
+
+- 6002feb3d: Fix exports in package.json files for react-native projects
+- Updated dependencies [6002feb3d]
+  - @graphql-codegen/visitor-plugin-common@2.5.2
+  - @graphql-codegen/plugin-helpers@2.3.2
+
 ## 1.2.0
 
 ### Minor Changes