+++ packages/presets/import-types/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @graphql-codegen/import-types-preset
 
+## 2.1.9
+
+### Patch Changes
+
+- 6002feb3d: Fix exports in package.json files for react-native projects
+- Updated dependencies [8643b3bf3]
+- Updated dependencies [6002feb3d]
+  - @graphql-codegen/add@3.1.1
+  - @graphql-codegen/visitor-plugin-common@2.5.2
+  - @graphql-codegen/plugin-helpers@2.3.2
+
 ## 2.1.8
 
 ### Patch Changes